+++ CHANGES.md
@@ -13,6 +13,8 @@ v0.12.4 (2017-01-09)
 * Fixed a bug that caused layers to not always be properly removed
   when deleting a row from the layer list. [#1502]
 
+* Make JSON circular reference errors more explicit. [#1529]
+
 v0.12.3 (2017-11-14)
 --------------------